annotate, rename and move enums and builders
