+++ DSharpPlus/Entities/AutoModeration/DiscordRuleEventType.cs
@@ -1,9 +1,9 @@
-namespace DSharpPlus.Enums;
+namespace DSharpPlus.Entities;
 
 /// <summary>
 /// Indicates in what event context a rule should be checked.
 /// </summary>
-public enum RuleEventType
+public enum DiscordRuleEventType
 {
     /// <summary>
     /// The rule will trigger when a member send or modify a message.
